requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
supabase>=2.10.0
python-dotenv>=1.0.1
pydantic[email]>=2.10.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.2
